Add DynamicLogic
