+++ app/Espo/Services/Record.php
@@ -39,7 +39,7 @@ use \Espo\Core\Exceptions\NotFound;
 use \Espo\Core\Utils\Util;
 use Espo\ORM\IEntity;
 use Treo\Core\Exceptions\NoChange;
-use Treo\Services\DynamicLogic;
+use Treo\Core\Utils\Condition\Condition;
 
 class Record extends \Espo\Core\Services\Base
 {
@@ -103,6 +103,11 @@ class Record extends \Espo\Core\Services\Base
     protected $mandatorySelectAttributeList = [];
 
     protected $forceSelectAllAttributes = false;
+
+    /**
+     * @var bool|array
+     */
+    private $relationFields = false;
 
     const MAX_SELECT_TEXT_ATTRIBUTE_LENGTH = 5000;
 
@@ -407,11 +412,8 @@ class Record extends \Espo\Core\Services\Base
      */
     protected function isValid($entity)
     {
-        /** @var DynamicLogic $dynamicLogic */
-        $dynamicLogic = $this->getServiceFactory()->create('DynamicLogic');
-
         foreach ($entity->getAttributes() as $field => $data) {
-            if ((!empty($data['required']) || $dynamicLogic->isRequiredField($field, $entity, 'required'))
+            if ((!empty($data['required']) || $this->isRequiredField($field, $entity, 'required'))
                 && is_null($entity->get($field))) {
                 throw new BadRequest("Validation failed. '$field' is required");
             }
@@ -2372,6 +2374,47 @@ class Record extends \Espo\Core\Services\Base
 
         if (($hasAssigned || $entity->hasAttribute('assignedUserId')) && empty($entity->get('assignedUserId'))) {
             $entity->set('assignedUserId', $this->getEntityManager()->getUser()->id);
+        }
+    }
+
+    /**
+     * @param string $field
+     * @param Entity $entity
+     * @param $typeResult
+     *
+     * @return bool
+     * @throws Error
+     */
+    public function isRequiredField(string $field, Entity $entity, $typeResult): bool
+    {
+        if ($this->relationFields === false) {
+            $this->setRelationFields($entity);
+        }
+        if (isset($this->relationFields[$field])) {
+            $field = $this->relationFields[$field];
+        }
+
+        $result = false;
+
+        $item = $this->getMetadata()
+            ->get("clientDefs.{$entity->getEntityName()}.dynamicLogic.fields.$field.$typeResult.conditionGroup", []);
+
+        if (!empty($item)) {
+            $result = Condition::isCheck(Condition::prepare($entity, $item));
+        }
+
+        return $result;
+    }
+
+    /**
+     * @param Entity $entity
+     */
+    private function setRelationFields(Entity $entity): void
+    {
+        foreach ($entity->getRelations() as $key => $relation) {
+            if (isset($relation['key'])) {
+                $this->relationFields[$relation['key']] = $key;
+            }
         }
     }
 }
